Update diversity.R
Thanks for this fantastic tool! I was trying out your clonalDiversity function as follows:

divDF <- clonalDiversity(combined, cloneCall = "collapsedClone", group = "cell_type", n.boots=100, exportTable = TRUE)

However, when examining the table, I noticed the row names did not match the values in the "cell_type" column. Based on my own calculations, it appears the row names are accurate while the cell_type column is not. I've added a few "sorts" to this portion of the code to resolve this discrepancy.